requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-bootstrap-modal-forms==2.2.0
django-heroku==0.0.0
django-tempus-dominus==5.1.2.14
django-widget-tweaks==1.4.8
gunicorn==20.1.0
idna==3.2
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
